requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.2
motor==3.4.0
yt-dlp==2024.5.27
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-adk==0.1.0
google-generativeai==0.7.0
streamlit==1.36.0
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.4
requests==2.32.3
isodate==0.6.1
pytz==2024.1
pandas==2.2.2
plotly==5.22.0
functions-framework==3.8.0
gunicorn==22.0.0